+++ src/kamihi/bot/bot.py
@@ -30,12 +30,13 @@ from telegram import BotCommand
 from telegram.ext import CommandHandler
 
 from kamihi.base.config import KamihiSettings
-from kamihi.bot.action import Action
 from kamihi.templates import Templates
 from kamihi.tg import TelegramClient
-from kamihi.users.models.user import User
-from kamihi.users.users import get_users, is_user_authorized
-from kamihi.web.web import KamihiWeb
+from kamihi.users import get_users, is_user_authorized
+from kamihi.users.models import User
+from kamihi.web import KamihiWeb
+
+from .action import Action
 
 
 class Bot:
