+++ src/kamihi/tg/client.py
@@ -32,7 +32,8 @@ from telegram.ext import (
 )
 
 from kamihi.base.config import KamihiSettings
-from kamihi.tg.default_handlers import default, error
+
+from .default_handlers import default, error
 
 
 async def _post_init(_: Application) -> None:
