+++ src/kamihi/tg/default_handlers.py
@@ -10,7 +10,7 @@ from loguru import logger
 from telegram import Update
 from telegram.ext import CallbackContext
 
-from kamihi.tg.send import reply_text
+from .send import reply_text
 
 
 async def default(update: Update, context: CallbackContext) -> None:
