+++ src/kamihi/users/models/__init__.py
@@ -15,7 +15,8 @@ Attributes:
 
 """
 
+from .permission import Permission
 from .role import Role
 from .user import User
 
-__all__ = ["User", "Role"]
+__all__ = ["User", "Role", "Permission"]
