+++ src/kamihi/web/__init__.py
@@ -5,3 +5,7 @@ License:
     MIT
 
 """
+
+from .web import KamihiWeb
+
+__all__ = ["KamihiWeb"]
